requirements.txt
streamlit
google-generativeai
pandas
numpy
networkx
python-Levenshtein
thefuzz